hw5/requirements-tests.txt
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
confluent-kafka==2.4.0
clickhouse-connect==0.7.16
psycopg2-binary==2.9.9
minio==7.2.7
protobuf==5.27.1
grpcio-tools==1.64.1
pydantic==2.7.4
requests==2.32.3
